+++ catalog/resource_storage_credential.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/databricks/databricks-sdk-go"
 	"github.com/databricks/databricks-sdk-go/service/catalog"
+	"github.com/databricks/terraform-provider-databricks/catalog/bindings"
 	"github.com/databricks/terraform-provider-databricks/common"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
@@ -21,6 +22,7 @@ type StorageCredentialInfo struct {
 	MetastoreID                 string                                       `json:"metastore_id,omitempty" tf:"computed"`
 	ReadOnly                    bool                                         `json:"read_only,omitempty"`
 	SkipValidation              bool                                         `json:"skip_validation,omitempty"`
+	IsolationMode               string                                       `json:"isolation_mode,omitempty" tf:"computed"`
 }
 
 func removeGcpSaField(originalSchema map[string]*schema.Schema) map[string]*schema.Schema {
@@ -71,10 +73,11 @@ func ResourceStorageCredential() common.Resource {
 				}
 				d.SetId(storageCredential.CredentialInfo.Name)
 
-				// Don't update owner if it is not provided
-				if d.Get("owner") == "" {
+				// Update owner or isolation mode if it is provided
+				if !updateRequired(d, []string{"owner", "isolation_mode"}) {
 					return nil
 				}
+
 				update.Name = d.Id()
 				_, err = acc.StorageCredentials.Update(ctx, catalog.AccountsUpdateStorageCredential{
 					CredentialInfo:        &update,
@@ -96,8 +99,8 @@ func ResourceStorageCredential() common.Resource {
 				}
 				d.SetId(storageCredential.Name)
 
-				// Don't update owner if it is not provided
-				if d.Get("owner") == "" {
+				// Update owner or isolation mode if it is provided
+				if !updateRequired(d, []string{"owner", "isolation_mode"}) {
 					return nil
 				}
 
@@ -106,7 +109,8 @@ func ResourceStorageCredential() common.Resource {
 				if err != nil {
 					return err
 				}
-				return nil
+				// Bind the current workspace if the storage credential is isolated, otherwise the read will fail
+				return bindings.AddCurrentWorkspaceBindings(ctx, d, w, storageCredential.Name, "storage-credential")
 			})
 		},
 		Read: func(ctx context.Context, d *schema.ResourceData, c *common.DatabricksClient) error {
@@ -241,7 +245,8 @@ func ResourceStorageCredential() common.Resource {
 					}
 					return err
 				}
-				return nil
+				// Bind the current workspace if the storage credential is isolated, otherwise the read will fail
+				return bindings.AddCurrentWorkspaceBindings(ctx, d, w, update.Name, "storage-credential")
 			})
 		},
 		Delete: func(ctx context.Context, d *schema.ResourceData, c *common.DatabricksClient) error {